arm64: bug: Add reachable annotation to warning macros
WARN* and BUG* both use brk #0x800 opcodes and the distinction is
provided by the contents of the bug table. This table is not accessible
to objtool, so add an annotation to WARN* macros to let objtool know
that brk handler will return an resume the execution of the instructions
following the WARN's brk.

Signed-off-by: Julien Thierry <jthierry@redhat.com>
[ 6.1 fp: annotate_reachable replaced by ASM_REACHABLE ]
Signed-off-by: Suraj Jitindar Singh <surajjs@amazon.com>