linux/kvm.h: Fix KVM_CAP_PTP_KVM numbering to match upstream
When this patch went upstream the KVM_CAP_PTP_KVM number was 195, it was
backported as 191, and it finally ended up being 198 upstream. This will cause
issues if a guest uses a different number to the host as incorrect
functionality will result.

Update it to be 198 to match upstream to avoid this issue.

Signed-off-by: Suraj Jitindar Singh <surajjs@amazon.com>