arm64,rethook,kprobes: Replace kretprobe with rethook on arm64
[ NOT UPSTREAM ]

In a way, this is similar to:

commit f3a112c0c40d ("x86,rethook,kprobes: Replace kretprobe with rethook on x86")

Except that, on arm64, the layout of the ftrace_regs structure does not
mirror pt_regs so by replacing kretprobe with rethook we actually have
kretprobe save less registers than it used to. In order not to break the
API, kretprobe internally reconstructs a sparse pt_regs it hands over to
kretprobe callbacks. Some of the fields which are no longer saved by the
trampoline may therefore be invalid.

Signed-off-by: Florent Revest <revest@chromium.org>