bpf: Fix bpf get_func_ip() on arm64 multi-kprobe programs
[ NOT UPSTREAM ]

Now that fprobe is implemented on arm64, "multi kprobe" BPF programs
attach but the entry_ip they get from ftrace is the address of the BL
instruction in the ftrace patch site so it can not be compared to the
address of a symbol, like in the multi_kprobe BPF selftests.

x86 already has a similar need for a fixup routine when IBT is used, so
there's already a get_entry_ip indirection to take this into account but
this macro is applied indistinctively for both kprobe and fprobe entry
addresses. On arm64, these addresses have different offsets so the macro
is split in two.

Signed-off-by: Florent Revest <revest@chromium.org>