Revert "x86/cpu: Actually turn off mitigations by default for SPECULATION_MITIGATIONS=n"
This reverts commit 9c09773917fbb77dff85b433e1e89123fc5fb530 merged in
v6.1.87.

Original commit turned off mitigations regardless of any architecutre if
CONFIG_SPECULATION_MITIGATIONS is disabled and this config is only defined in x86_64.
Therefore mitigations in arm64 are turned off as well. Fixes were
proposed by commit author in
https://lore.kernel.org/all/20240417001507.2264512-1-seanjc@google.com/#t
but not merged to mainline yet. AL kernel rebased to v6.1.87 will revert this
commit to unblock the release.

Signed-off-by: Shaoying Xu <shaoyi@amazon.com>